refactor!: prefer `ExportedBlock` over `Block`
Since the proof is not transferred from the node, the block hash cannot
be calculated. For this reason, we rely on the upstream data structure
`ExportedBlock` instead of `Block`, as that allows for the hash
calculation without access to the proof.

It also have more helper functions required for transaction handling,
such as methods related to the mutator set.